Filter views: store views regardless of the persistence of the filters
Change-type: patch
